Fix issue #12 - Colours are now unique
Implements a fix to ensure colours are unique. In addition it
implements ideas proposed by @buffcode such that at the end of the
process we sort the variable colorCounts and return the
1..<paletteSize> items (thus dropping the dominant color). The
implementation of the new helper functions and return values also make
it easy for others should they wish to amend the palette’s behaviour.