Fix issue #8 - 1 & 2 of 3 - Stop skipping over 84%  of all pixels
Removes blocksize iteration skips over 4/5 of all pixels which is not
desirable.

Removes constraint preventing to go beyond the 1st quarter of the
image. The ( img.width * img.height ) calculation is 4 times shorter
than data.length and consistently gets the upper hand in the evaluation
to set the number of pixels that get analysed meaning we’re left with
only the upper quarter of the image analysed.