Fix issue #33 - Set canvas to image size instead of  300 by 150
When RGBaster instantiates the canvas we now pass it the image size to
ensure the canvas and image size are the same. The issue was that with
no canvas size set, it defaulted to 300 by 150. This caused issues as
the image could be larger than the canvas and thus only partially
painted onto the canvas. Reading back from the canvas was equally an
issue as portions of the image would be unavailable.